fixes
re-enable file locks, found they are necessary for writes that involve multiple chunks.  Write fails in Zarr result in NaN on disk.
with vertical reference ellipse, if the input/output datum do not match, we perform a vertical transformation
working on custom grid datum where we can apply a gridded value to soundings.